requirements.txt
Django==1.4.2
MySQL-python==1.2.4c1
PIL==1.1.7
Wave==0.0.2
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.17
celery==3.0.11
django-celery==3.0.11
django-facebook==4.2.2
django-picklefield==0.2.1
django-tastypie==0.9.11
django-templateaddons==0.1
kombu==2.4.8
mimeparse==0.1.3
numpy==1.6.2
pyechonest==4.3.0
python-dateutil==1.5
pytz==2012h
scikits.audiolab==0.11.0
wsgiref==0.1.2
